Move the region switcher out of the secondary nav and into the gutter when the gutter is collapsed
