+++ ui/app/styles/core/navbar.scss
@@ -87,6 +87,7 @@
 
       @media #{$mq-hidden-gutter} {
         width: 0;
+        visibility: hidden;
       }
     }
   }
